Fixed up some of the build settings.
I also added a couple of submodules to make building a bit easier.  However
I removed ffmpeg as a submodule.  After messing with it a bit I decided that it's
not currently worth the effort to include it.  Homebrew, or macports is probably
more suitable.  In the future it may be included, but I think that will be if, and when
ffmpeg as a library is used in a more integrated manner.