Modified the preferences quite a bit.
I updated this to use MASPreferences.  Which is mostly a personal preference of mine,
but it did seem to reduce the amount of code.  I've also switched to using NSUserDefaults
rather than tracking this stuff ourselves.

I also removed all the bundled.  It's now just using the main bundle.  If needed they can
be added back later, but for the time being they seemed a little premature.